feat: add exams table

+++ src/model/entities/exames/tipos/Eletrocardiograma.java
@@ -6,7 +6,7 @@ import model.entities.exames.Exame;
 public class Eletrocardiograma extends BaseEntity {
 	String ritmo;
 	int fc;
-	Double ondaP;
+	float ondaP;
 	int complexoQRS;
 	int eixoEletrico;
 	Exame exame;
